Address some review comments
