+++ .lxpreferences
@@ -1,8 +1,8 @@
 {
   "version": "0.4.1",
   "projectFileName": "Vehicle.lxp",
-  "windwWidth": 1792,
-  "windowHeight": 983,
+  "windwWidth": 1728,
+  "windowHeight": 984,
   "uiZoom": 100,
   "focusChannelOnCue": false,
   "focusActivePattern": false,
