+++ src/main/java/titanicsend/app/autopilot/TEShowKontrol.java
@@ -1,0 +1,8 @@
+package titanicsend.app.autopilot;
+
+/**
+ * Holds configuration around how TE connects to ShowKontrol
+ */
+public class TEShowKontrol {
+    public static final int OSC_PORT = 42069;
+}
